Export BinaryDataType enum
