+++ src/constants/Constants.ts
@@ -12,3 +12,11 @@ export enum RPCResponseSubject {
   PersonUpdate = 'person_update',
   PersonsAlive = 'persons_alive',
 }
+
+export enum BinaryDataType {
+  TYPE_IMAGE = 1,
+  TYPE_SKELETON = 2,
+  TYPE_THUMBNAIL = 3,
+  TYPE_HEATMAP = 4,
+  TYPE_DEPTHMAP = 5,
+}
